+++ .gitignore
@@ -1,1 +1,1 @@
-/gitrepo-version
+/gitsemver
