+++ pkg/gitsemver/validate.go
@@ -1,4 +1,4 @@
-package gitrepo
+package gitsemver
 
 import "regexp"
