+++ pkg/gitsemver/validate_test.go
@@ -1,4 +1,4 @@
-package gitrepo
+package gitsemver
 
 import "testing"
